feat: added format subdirective to commodity directive content line ruleset
This change in parsing rule is required to preserve the order in which the format subdirective
appears among subdirective comments. Previously, the format subdirective was treated separately,
which removes information of its placement among comments. The drawback of this approach is that the
parser is unable to enforce unique format subdirectives, which need to be handled elsewhere.